backend/requirements.txt
websocket-client==1.8.0
streamcontroller-plugin-tools==2.0.2
